Don't add TextGeometry with no text the separator has label '', if you try to instanciate a TextGeometry with this it results in this error:
`[.CommandBufferContext]RENDER WARNING: Render count or primcount is 0.`

To fix the problem I check for a separator in the array before trying to create TextGeometry.  If a separator is found I add an empty Object3D so that the items group has the correct length, thus allowing the selector's position to not be fucked up.